fix no NIOSSL build
